requirements.txt
kafka-python==1.4.4
kazoo==2.6.1
pure-sasl==0.5.1
